+++ 99.npm/main.js
@@ -1,85 +1,17 @@
 import express from "express";
 import open from "open";
-import enquirer from "enquirer";
 import puppeteer from "puppeteer";
-import { cities } from "./cities.js";
-import { months } from "./months.js";
-import { SurveyPoint } from "./surveyPoint.js";
+import { SurveyCLI } from "./surveyCLI.js";
 
-const { Select } = enquirer;
-const minYear = 1976;
 const surveyDataStartIndex = 22;
 const surveyDataEndIndex = -4;
 
 const main = async () => {
-  const promptPrefecture = new Select({
-    name: "value",
-    message: "都道府県または地方を選んでください。",
-    choices: Object.keys(cities),
-  });
-
-  const prefecture = await promptPrefecture.run();
-
-  const promptCity = new Select({
-    name: "value",
-    message: "市町村を選んでください。",
-    choices: Object.keys(cities[prefecture]),
-  });
-
-  const city = await promptCity.run();
-  const baseYears = [];
-  for (let year = new Date().getFullYear(); year >= minYear + 1; year--) {
-    baseYears.push(`${year} 年`);
-  }
-
-  const promptBaseYear = new Select({
-    name: "value",
-    message: "基準年を選んでください。",
-    choices: baseYears,
-  });
-
-  const baseYear = (await promptBaseYear.run()).replace(/[^0-9]/g, "");
-  console.log(typeof baseYear);
-
-  const targetYears = [];
-  for (let year = baseYear - 1; year >= minYear; year--) {
-    targetYears.push(`${year} 年`);
-  }
-
-  const promptTargetYear = new Select({
-    name: "value",
-    message: "比較年を選んでください。",
-    choices: targetYears,
-  });
-
-  const targetYear = (await promptTargetYear.run()).replace(/[^0-9]/g, "");
-
-  const promptMonth = new Select({
-    name: "value",
-    message: "基準月を選んでください。",
-    choices: months,
-  });
-
-  const month = (await promptMonth.run()).replace(/[^0-9]/g, "");
+  const surveyCLI = new SurveyCLI();
+  const [baseSurveyPoint, targetSurveyPoint, city] = await surveyCLI.execute();
 
   const browser = await puppeteer.launch();
   const page = await browser.newPage();
-
-  const precNo = cities[prefecture][city].precNo;
-  const blockNo = cities[prefecture][city].blockNo;
-
-  const baseSurveyPoint = new SurveyPoint({
-    precNo,
-    blockNo,
-    year: baseYear,
-    month,
-  });
-  const targetSurveyPoint = new SurveyPoint({
-    precNo,
-    blockNo,
-    year: targetYear,
-    month,
-  });
 
   await page.goto(baseSurveyPoint.url());
 
@@ -192,8 +124,12 @@ const main = async () => {
                 10,
             ) / 10
           } ℃";
-          const baseLabel = "${baseYear} 年 ${month} 月";
-          const targetLabel = "${targetYear} 年 ${month} 月";
+          const baseLabel = "${baseSurveyPoint.year} 年 ${
+            baseSurveyPoint.month
+          } 月";
+          const targetLabel = "${targetSurveyPoint.year} 年 ${
+            targetSurveyPoint.month
+          } 月";
           document.getElementById("title").textContent = titleText;
           document.getElementById("average").textContent = averageText;
           document.getElementById("highest").textContent = highestText;
@@ -245,7 +181,7 @@ const main = async () => {
 
   const server = app.listen(port, () => {
     console.log(`Server is runnnin on port ${port}`);
-    console.log(`"Please press Ctrl + C to exit."`);
+    console.log(`"This will exit in 10 seconds."`);
   });
 
   open(`http://localhost:${port}`);
